+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 
 ### Breaking
 
+- [#5559](https://github.com/ChainSafe/forest/pull/5559) Change `Filecoin.ChainGetMinBaseFee` to `Forest.ChainGetMinBaseFee` with read access.
+
 ### Added
 
 ### Changed